Börjat dela upp customer
