+++ CustomerAddUser.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DotNetDynamos
+{
+    internal partial class Customer : AllUsers
+    {
+        public void AddUser(string username, string password, int IDnumber)
+        {
+            _username = username;
+            Password = password;
+            _IDnumber = IDnumber;
+            CustomerUsers.Add(IDnumber, password);
+        }
+    }
+}
